Generate Garden project with Dev Mode
Dev Mode is configured to use Flask for the development loop. This does
not currently function as expected as Flask will give an error on change
of the Hello, World string in `app.py` claiming that app.py does not
exist.

To test, deploy the app to a local cluster with `garden deploy --dev
frontend`, change the "Hello, World!" return to "Hello, Mom!" and reload
the ingress endpoint.